feat: added data-fetching story, refactored Link out of stories
