+++ starters/qwik-fetch-tailwind/src/routes/counter/index.tsx
@@ -1,6 +1,12 @@
 import { component$ } from '@builder.io/qwik';
+import { Link } from '@builder.io/qwik-city';
 import { Counter } from './counter';
 
 export default component$(() => {
-  return <Counter />;
+  return (
+    <div>
+      <Counter />
+      <Link href="/">Return Home</Link>
+    </div>
+  );
 });
