+++ starters/qwik-fetch-tailwind/src/routes/data-fetching/data-fetching.stories.tsx
@@ -1,0 +1,10 @@
+import { Meta } from '@storybook/html';
+import { DataFetching } from './data-fetching';
+
+export default {
+  title: 'Data Fetching',
+} as Meta;
+
+const Template = () => <DataFetching />;
+
+export const Demo = Template.bind({});
